Refactor database table creation, add README
Improved functionality of the table creation method in 'myApp.cpp' by creating indexes on 'name' and 'gender' attributes for faster SELECT queries. This was necessary to optimize the application's data manipulation and retrieval speed.
Also, added a README.md file to provide clear instructions and information on how to install and use the application.