Update CMakeLists and README file
Lowered the minimum required CMake version and CXX standard in the CMakeLists file for broader compatibility. This was achieved by changing the minimum required version from 3.26 to 3.16 and the CXX standard from 17 to 11.
The README file was updated with more comprehensive instructions for compilation and clarified usage details. Added a step to input the SQLiteCpp path during compilation, and clarified the parameters that the application accepts.